fix: 로그인 로그 및 상태코드 상속 문제 개선#12
